fix(hub): move cross-replica test behind !no_sqlite build tag
TestChatNotificationSubject_CrossReplicaChannel uses
newTestPostgresPublisher, which is defined in events_postgres_test.go
behind //go:build !no_sqlite. The test lived in
chat_notification_scope_test.go (no build tag), so `go vet -tags
no_sqlite` failed with "undefined: newTestPostgresPublisher".

Move the test to its own file with the matching build constraint.
